fix(eslint): runtime error for empty modules (#1754)
* test(eslint): empty module should be valid in lexical scope rule

Add a test to verify that an empty module does not report any violations
and has no runtime errors during linting.

The test fails at the moment. There is a runtime error when linting
empty files. The error will be fixed in a follow-up commit.

* fix(eslint): check if moduleSymbol exists before asking for exports

The `typeChecker.getSymbolAtLocation` method returns `undefined` for
empty modules, despite what the type signature of that method says.

Passing that `undefined` value to `typeChecker.getExportsOfModule` leads
to a runtime error (trying to access `undefined.flags`). This error was
caught by the test added in the previous commit.

Only calling `typeChecker.getExportsOfModule` conditionally fixes the
error.